Resolve BASIS-175051

Closes BASIS-175051

See merge request basis/abaco_codigo_fonte!88

+++ frontend/src/app/funcao-transacao/funcao-transacao-divergence/funcao-transacao-divergence.component.ts
@@ -845,32 +845,53 @@ export class FuncaoTransacaoDivergenceComponent implements OnInit {
         this.pageNotificationService.addSuccessMessage('Status da(s) funcionalidade(s) alterado(s).');
     }
 
+    confirmDelete(funcaoTransacaoSelecionada: FuncaoTransacao) {
+        this.confirmationService.confirm({
+            message: `${
+                this.getLabel('Tem certeza que deseja alterar o status para Excluido a Função de Transação')
+            } '${funcaoTransacaoSelecionada.name}'?`,
+            accept: () => {
+                this.funcaoTransacaoService.deleteStatus(funcaoTransacaoSelecionada.id).subscribe(value => {
+                    funcaoTransacaoSelecionada = this.funcoesTransacoes.filter((funcaoTransacao) => (funcaoTransacao.id === funcaoTransacaoSelecionada.id))[0];
+                    funcaoTransacaoSelecionada['statusFuncao'] = value['statusFuncao'];
+                    this.pageNotificationService.addSuccessMessage('Status da funcionalidade ' + funcaoTransacaoSelecionada.name + ' foi alterado.');
+                    this.divergenciaService.updateDivergenciaSomaPf(this.analise.id).subscribe();
+                });
+            }
         });
-        this.pageNotificationService.addSuccessMessage('Status da(s) funcionalidade(s) alterado(s).');
-    } 
-
-    setApproved(funcaoTransacaoSelecionadas: FuncaoTransacao[]) {
-        funcaoTransacaoSelecionadas.forEach(funcaoTransacaoSelecionada => {
-            this.funcaoTransacaoService.approved(funcaoTransacaoSelecionada.id).subscribe(value => {
-                funcaoTransacaoSelecionada = this.funcoesTransacoes.filter((funcaoTransacao) => (funcaoTransacao.id === funcaoTransacaoSelecionada.id))[0];
-                funcaoTransacaoSelecionada['statusFuncao'] = value['statusFuncao'];
-            });
-            
-        });
-        this.pageNotificationService.addSuccessMessage('Status da(s) funcionalidade(s) alterado(s).');
-
     }
 
-    setDelete(funcaoTransacaoSelecionadas: FuncaoTransacao[]) {
-       
-        funcaoTransacaoSelecionadas.forEach(funcaoTransacaoSelecionada => {
-            this.funcaoTransacaoService.deleteStatus(funcaoTransacaoSelecionada.id).subscribe(value => {
-                funcaoTransacaoSelecionada = this.funcoesTransacoes.filter((funcaoTransacao) => (funcaoTransacao.id === funcaoTransacaoSelecionada.id))[0];
-                funcaoTransacaoSelecionada['statusFuncao'] = value['statusFuncao'];
-            });
 
+    confirmDivergence(funcaoTransacaoSelecionada: FuncaoTransacao) {
+        this.confirmationService.confirm({
+            message: `${
+                this.getLabel('Tem certeza que deseja alterar o status para Divergente a Função de Transação')
+            } '${funcaoTransacaoSelecionada.name}'?`,
+            accept: () => {
+                this.funcaoTransacaoService.pending(funcaoTransacaoSelecionada.id).subscribe(value => {
+                    funcaoTransacaoSelecionada = this.funcoesTransacoes.filter((funcaoTransacao) => (funcaoTransacao.id === funcaoTransacaoSelecionada.id))[0];
+                    funcaoTransacaoSelecionada['statusFuncao'] = value['statusFuncao'];
+                    this.pageNotificationService.addSuccessMessage('Status da funcionalidade ' + funcaoTransacaoSelecionada.name + ' foi alterado.');
+                    this.divergenciaService.updateDivergenciaSomaPf(this.analise.id).subscribe();
+                });
+            }
         });
-        this.pageNotificationService.addSuccessMessage('Status da(s) funcionalidade(s) alterado(s).');
+    }
+
+    confirmApproved(funcaoTransacaoSelecionada: FuncaoTransacao) {
+        this.confirmationService.confirm({
+            message: `${
+                this.getLabel('Tem certeza que deseja alterar o status para Aprovado a Função de Transação')
+            } '${funcaoTransacaoSelecionada.name}'?`,
+            accept: () => {
+                this.funcaoTransacaoService.approved(funcaoTransacaoSelecionada.id).subscribe(value => {
+                    funcaoTransacaoSelecionada = this.funcoesTransacoes.filter((funcaoTransacao) => (funcaoTransacao.id === funcaoTransacaoSelecionada.id))[0];
+                    funcaoTransacaoSelecionada['statusFuncao'] = value['statusFuncao'];
+                    this.pageNotificationService.addSuccessMessage('Status da funcionalidade ' + funcaoTransacaoSelecionada.name + ' foi alterado.');
+                    this.divergenciaService.updateDivergenciaSomaPf(this.analise.id).subscribe();
+                });
+            }
+        });
     }
 
     formataFatorAjuste(fatorAjuste: FatorAjuste): string {
